made corrections to assignment 1 and 4

+++ week01/Exercise4/Program.cs
@@ -9,7 +9,7 @@ class Program
         int total = 0;
         int items = 0;
         int largest = 0;
-        int average;
+        float average;
         Console.WriteLine("Enter a list of numbers, type 0 when finished.");
         while (newNumber != 0)
         {
@@ -29,7 +29,7 @@ class Program
 
             }
         }
-        average = total / items;
+        average = (float)total / items;
         Console.WriteLine($"The sum is: {total}");
         Console.WriteLine($"The average is: {average}");
         Console.WriteLine($"The largest number is:{largest}");
